Organize plugins into subfolders
