+++ src/main/scala/t800/plugins/execute/SecondaryInstrPlugin.scala
@@ -8,6 +8,8 @@ import spinal.core.fiber.Retainer
 import t800.{Opcodes, Global}
 import t800.plugins.{ChannelSrv, ChannelTxCmd, LinkBusSrv, LinkBusArbiterSrv}
 import t800.plugins.schedule.SchedSrv
+import t800.plugins.fpu.FpuSrv
+import t800.plugins.timers.TimerSrv
 import scala.util.Try
 
 /** Implements basic ALU instructions and connects to the global pipeline. */
