+++ src/main/scala/t800/plugins/grouper/Service.scala
@@ -1,0 +1,14 @@
+package t800.plugins.grouper
+
+import spinal.core._
+import spinal.lib._
+import t800.Global
+
+case class GroupedInstructions() extends Bundle {
+  val instructions = Vec(Bits(Global.OPCODE_BITS bits), 8)
+  val count = UInt(4 bits)
+}
+
+trait GroupedInstrSrv {
+  def groups: Flow[GroupedInstructions]
+}
