+++ src/main/scala/t800/plugins/timers/TimerPlugin.scala
@@ -1,4 +1,4 @@
-package t800.plugins
+package t800.plugins.timers
 
 import spinal.core._
 import spinal.lib._
@@ -6,6 +6,7 @@ import spinal.core.sim._
 import spinal.core.fiber.Retainer
 import spinal.lib.misc.plugin._
 import t800.Global
+import t800.plugins.timers._
 
 /** Simple high/low priority timers. High increments every cycle; low every 64 cycles. */
 class TimerPlugin extends FiberPlugin {
