+++ src/main/scala/t800/plugins/transputer/Service.scala
@@ -1,0 +1,4 @@
+package t800.plugins.transputer
+
+/** Placeholder for Transputer-specific services. */
+trait TransputerService
