+++ src/test/scala/t800/GrouperPluginSpec.scala
@@ -4,6 +4,7 @@ import spinal.core._
 import spinal.core.sim._
 import org.scalatest.funsuite.AnyFunSuite
 import t800.plugins._
+import t800.plugins.grouper.GrouperPlugin
 import spinal.lib.misc.plugin.PluginHost
 
 class GrouperPluginSpec extends AnyFunSuite {
