+++ src/test/scala/t800/HelloWorldSim.scala
@@ -5,6 +5,8 @@ import spinal.core.sim._
 import t800.plugins._
 import spinal.lib.misc.plugin.PluginHost
 import t800.plugins.schedule.SchedulerPlugin
+import t800.plugins.timers.TimerPlugin
+import t800.plugins.grouper.GrouperPlugin
 
 object HelloWorldSim {
