+++ src/test/scala/t800/HelloWorldSpec.scala
@@ -6,6 +6,8 @@ import org.scalatest.funsuite.AnyFunSuite
 import t800.plugins._
 import spinal.lib.misc.plugin.PluginHost
 import t800.plugins.schedule.SchedulerPlugin
+import t800.plugins.timers.TimerPlugin
+import t800.plugins.grouper.GrouperPlugin
 
 class HelloWorldSpec extends AnyFunSuite {
   ignore("ROM program prints hello world") {
