+++ src/test/scala/t800/Move2DSpec.scala
@@ -6,6 +6,7 @@ import org.scalatest.funsuite.AnyFunSuite
 import t800.plugins._
 import spinal.lib.misc.plugin.PluginHost
 import t800.plugins.schedule.SchedulerPlugin
+import t800.plugins.grouper.GrouperPlugin
 
 class Move2DSpec extends AnyFunSuite {
   test("MOVE2DALL transfers 2D bytes") {
