+++ src/test/scala/t800/OprStlSpec.scala
@@ -8,6 +8,7 @@ import spinal.lib._
 import t800.plugins.schedule.SchedulerPlugin
 import spinal.lib.misc.plugin.PluginHost
 import t800.{DummyTimerPlugin, DummyFpuPlugin}
+import t800.plugins.grouper.GrouperPlugin
 
 class OprStlSpec extends AnyFunSuite {
   test("STL stores A and updates stack") {
