+++ src/test/scala/t800/PrefixSpec.scala
@@ -7,6 +7,7 @@ import t800.plugins._
 import spinal.lib.misc.plugin.PluginHost
 import t800.plugins.schedule.SchedulerPlugin
 import t800.{DummyTimerPlugin, DummyFpuPlugin}
+import t800.plugins.grouper.GrouperPlugin
 
 class PrefixSpec extends AnyFunSuite {
   test("PFIX/NFIX build literals") {
